refactor: fix name of directory
