+++ src/entity/application/usecases/delete.usecase.ts
