+++ src/entity/application/usecases/find-by-email.usecase.ts
