+++ src/entity/application/usecases/find-by-id.usecase.ts
