+++ src/entity/application/usecases/list.usecase.ts
