+++ src/entity/infrastructure/handlers/entity.handler.ts
@@ -4,13 +4,13 @@ import { Roles } from '../../../auth/domain/entities/role'
 import type IJWTProvider from '../../../auth/domain/providers/jwt.interface.provider'
 import HandleHTTPResponse from '../../../shared/utils/http.reply.util'
 import { GetPaginationParams, GetURLParams } from '../../../shared/utils/http.request.util'
-import DeleteEntityUseCase from '../../aplication/usecases/delete.usecase'
-import FindByEmailUseCase from '../../aplication/usecases/find-by-email.usecase'
-import FindEntityByIDUseCase from '../../aplication/usecases/find-by-id.usecase'
-import ListEntitiesUseCase from '../../aplication/usecases/list.usecase'
-import LoginEntityUseCase from '../../aplication/usecases/login.usecase'
-import RegisterEntityUseCase from '../../aplication/usecases/register.usecase'
-import UpdateEntityUseCase from '../../aplication/usecases/update.usecase'
+import DeleteEntityUseCase from '../../application/usecases/delete.usecase'
+import FindByEmailUseCase from '../../application/usecases/find-by-email.usecase'
+import FindEntityByIDUseCase from '../../application/usecases/find-by-id.usecase'
+import ListEntitiesUseCase from '../../application/usecases/list.usecase'
+import LoginEntityUseCase from '../../application/usecases/login.usecase'
+import RegisterEntityUseCase from '../../application/usecases/register.usecase'
+import UpdateEntityUseCase from '../../application/usecases/update.usecase'
 import type EntityLoginPayload from '../../domain/payloads/entity.login.payload'
 import type EntityPayload from '../../domain/payloads/entity.payload'
 import type EntityRepository from '../../domain/repositories/entity.repository'
