+++ src/green-point/application/usecases/delete.usecase.ts
