+++ src/green-point/application/usecases/find-by-id.usecase.ts
