+++ src/green-point/application/usecases/find-by-name.usecase.ts
