+++ src/green-point/infrastructure/handlers/green-point.handler.ts
@@ -1,13 +1,13 @@
 import { type FastifyReply, type FastifyRequest } from 'fastify'
-import FindEntityByIDUseCase from '../../../entity/aplication/usecases/find-by-id.usecase'
+import FindEntityByIDUseCase from '../../../entity/application/usecases/find-by-id.usecase'
 import type EntityRepository from '../../../entity/domain/repositories/entity.repository'
 import HandleHTTPResponse from '../../../shared/utils/http.reply.util'
 import { GetPaginationParams, GetURLParams } from '../../../shared/utils/http.request.util'
-import DeleteGreenPointUseCase from '../../aplication/usecases/delete.usecase'
-import FindGreenPointByIDUseCase from '../../aplication/usecases/find-by-id.usecase'
-import ListGreenPointsUseCase from '../../aplication/usecases/list.usecase'
-import RegisterGreenPointUseCase from '../../aplication/usecases/register.usecase'
-import UpdateGreenPointUseCase from '../../aplication/usecases/update.usecase'
+import DeleteGreenPointUseCase from '../../application/usecases/delete.usecase'
+import FindGreenPointByIDUseCase from '../../application/usecases/find-by-id.usecase'
+import ListGreenPointsUseCase from '../../application/usecases/list.usecase'
+import RegisterGreenPointUseCase from '../../application/usecases/register.usecase'
+import UpdateGreenPointUseCase from '../../application/usecases/update.usecase'
 import type GreenPointPayload from '../../domain/payloads/green-point.payload'
 import type GreenPointRepository from '../../domain/repositories/green-point.repository'
 import CheckIdDTO from '../dtos/check-id.dto'
