+++ src/neighbor/application/usecases/delete.usecase.ts
