+++ src/neighbor/application/usecases/find-by-email.usecase.ts
