+++ src/neighbor/application/usecases/find-by-id.usecase.ts
