+++ src/neighbor/infrastructure/handlers/neighbor.handler.ts
@@ -2,19 +2,19 @@ import { type FastifyReply, type FastifyRequest } from 'fastify'
 import AuthService from '../../../auth/aplicaction/service/auth.service'
 import { Roles } from '../../../auth/domain/entities/role'
 import type IJWTProvider from '../../../auth/domain/providers/jwt.interface.provider'
-import FindEntityByIDUseCase from '../../../entity/aplication/usecases/find-by-id.usecase'
+import FindEntityByIDUseCase from '../../../entity/application/usecases/find-by-id.usecase'
 import type EntityRepository from '../../../entity/domain/repositories/entity.repository'
 import DateUtils from '../../../shared/utils/date.util'
 import HandleHTTPResponse from '../../../shared/utils/http.reply.util'
 import { GetPaginationParams, GetURLParams } from '../../../shared/utils/http.request.util'
-import DeleteNeighborUseCase from '../../aplication/usecases/delete.usecase'
-import FindNeighborByDNIUseCase from '../../aplication/usecases/find-by-dni.usecase'
-import FindByEmailUseCase from '../../aplication/usecases/find-by-email.usecase'
-import FindNeighborByIDUseCase from '../../aplication/usecases/find-by-id.usecase'
-import ListNeighborsUseCase from '../../aplication/usecases/list.usecase'
-import LoginNeighborUseCase from '../../aplication/usecases/login.usecase'
-import RegisterNeighborUseCase from '../../aplication/usecases/register.usecase'
-import UpdateNeighborUseCase from '../../aplication/usecases/update.usecase'
+import DeleteNeighborUseCase from '../../application/usecases/delete.usecase'
+import FindNeighborByDNIUseCase from '../../application/usecases/find-by-dni.usecase'
+import FindByEmailUseCase from '../../application/usecases/find-by-email.usecase'
+import FindNeighborByIDUseCase from '../../application/usecases/find-by-id.usecase'
+import ListNeighborsUseCase from '../../application/usecases/list.usecase'
+import LoginNeighborUseCase from '../../application/usecases/login.usecase'
+import RegisterNeighborUseCase from '../../application/usecases/register.usecase'
+import UpdateNeighborUseCase from '../../application/usecases/update.usecase'
 import type NeighborPayload from '../../domain/payloads/neighbor.payload'
 import type NeighborRepository from '../../domain/repositories/neighbor.repository'
 import CheckIdDTO from '../dtos/check-id.dto'
