+++ src/responsible/application/usecases/delete.usecase.ts
