+++ src/responsible/application/usecases/find-by-email.usecase.ts
