+++ src/responsible/application/usecases/find-by-id.usecase.ts
