+++ src/responsible/application/usecases/list.usecase.ts
