+++ src/responsible/infrastructure/handlers/responsible.handler.ts
@@ -2,17 +2,17 @@ import { type FastifyReply, type FastifyRequest } from 'fastify'
 import AuthService from '../../../auth/aplicaction/service/auth.service'
 import { Roles } from '../../../auth/domain/entities/role'
 import type IJWTProvider from '../../../auth/domain/providers/jwt.interface.provider'
-import FindEntityByIDUseCase from '../../../entity/aplication/usecases/find-by-id.usecase'
+import FindEntityByIDUseCase from '../../../entity/application/usecases/find-by-id.usecase'
 import type EntityRepository from '../../../entity/domain/repositories/entity.repository'
 import HandleHTTPResponse from '../../../shared/utils/http.reply.util'
 import { GetPaginationParams, GetURLParams } from '../../../shared/utils/http.request.util'
-import DeleteResponsibleUseCase from '../../aplication/usecases/delete.usecase'
-import FindByEmailUseCase from '../../aplication/usecases/find-by-email.usecase'
-import FindResponsibleByIDUseCase from '../../aplication/usecases/find-by-id.usecase'
-import ListResponsiblesUseCase from '../../aplication/usecases/list.usecase'
-import LoginResponsibleUseCase from '../../aplication/usecases/login.usecase'
-import RegisterResponsibleUseCase from '../../aplication/usecases/register.usecase'
-import UpdateResponsibleUseCase from '../../aplication/usecases/update.usecase'
+import DeleteResponsibleUseCase from '../../application/usecases/delete.usecase'
+import FindByEmailUseCase from '../../application/usecases/find-by-email.usecase'
+import FindResponsibleByIDUseCase from '../../application/usecases/find-by-id.usecase'
+import ListResponsiblesUseCase from '../../application/usecases/list.usecase'
+import LoginResponsibleUseCase from '../../application/usecases/login.usecase'
+import RegisterResponsibleUseCase from '../../application/usecases/register.usecase'
+import UpdateResponsibleUseCase from '../../application/usecases/update.usecase'
 import type ResponsiblePayload from '../../domain/payloads/responsible.payload'
 import type ResponsibleRepository from '../../domain/repositories/responsible.repository'
 import CheckIdDTO from '../dtos/check-id.dto'
