+++ src/reward-partner/application/usecases/find-by-email.usecase.ts
