+++ src/reward-partner/infrastructure/handlers/reward-partner.handler.ts
@@ -3,14 +3,14 @@ import { type FastifyReply, type FastifyRequest } from 'fastify'
 import AuthService from '../../../auth/aplicaction/service/auth.service'
 import { Roles } from '../../../auth/domain/entities/role'
 import type IJWTProvider from '../../../auth/domain/providers/jwt.interface.provider'
-import FindEntityByIDUseCase from '../../../entity/aplication/usecases/find-by-id.usecase'
+import FindEntityByIDUseCase from '../../../entity/application/usecases/find-by-id.usecase'
 import type EntityRepository from '../../../entity/domain/repositories/entity.repository'
 import HandleHTTPResponse from '../../../shared/utils/http.reply.util'
 import { GetURLParams } from '../../../shared/utils/http.request.util'
-import FindByEmailUseCase from '../../aplication/usecases/find-by-email.usecase'
-import LoginRewardPartnerUseCase from '../../aplication/usecases/login.usecase'
-import RegisterRewardPartnerUseCase from '../../aplication/usecases/register.usecase'
-import UpdateRewardPartnerUseCase from '../../aplication/usecases/update.usecase'
+import FindByEmailUseCase from '../../application/usecases/find-by-email.usecase'
+import LoginRewardPartnerUseCase from '../../application/usecases/login.usecase'
+import RegisterRewardPartnerUseCase from '../../application/usecases/register.usecase'
+import UpdateRewardPartnerUseCase from '../../application/usecases/update.usecase'
 import type RewardPartnerLoginPayload from '../../domain/payloads/reward-partner.login.payload'
 import type RewardPartnerPayload from '../../domain/payloads/reward-partner.payload'
 import type RewardPartnerRepository from '../../domain/repositories/reward-partner.repository'
