+++ src/waste-category/application/usecases/delete.usecase.ts
