+++ src/waste-category/application/usecases/find-by-id.usecase.ts
