+++ src/waste-category/application/usecases/list.usecase.ts
