+++ src/waste-category/application/usecases/register.usecase.ts
