+++ src/waste-category/infrastructure/handlers/waste-category.handler.ts
@@ -1,11 +1,11 @@
 import { type FastifyReply, type FastifyRequest } from 'fastify'
 import HandleHTTPResponse from '../../../shared/utils/http.reply.util'
 import { GetPaginationParams, GetURLParams } from '../../../shared/utils/http.request.util'
-import DeleteCategoryUseCase from '../../aplication/usecases/delete.usecase'
-import FindWasteCategoryByIDUseCase from '../../aplication/usecases/find-by-id.usecase'
-import ListCategoriesUseCase from '../../aplication/usecases/list.usecase'
-import RegisterWasteCategoryUseCase from '../../aplication/usecases/register.usecase'
-import UpdateCategoryUseCase from '../../aplication/usecases/update.usecase'
+import DeleteCategoryUseCase from '../../application/usecases/delete.usecase'
+import FindWasteCategoryByIDUseCase from '../../application/usecases/find-by-id.usecase'
+import ListCategoriesUseCase from '../../application/usecases/list.usecase'
+import RegisterWasteCategoryUseCase from '../../application/usecases/register.usecase'
+import UpdateCategoryUseCase from '../../application/usecases/update.usecase'
 import type WasteCategoryPayload from '../../domain/payloads/waste-category.payload'
 import type WasteCategoryRepository from '../../domain/repositories/waste-category.repository'
 import CheckIdDTO from '../dtos/check-id.dto'
