+++ src/waste-transaction/infrastructure/handlers/waste-transaction.handler.ts
@@ -1,10 +1,10 @@
 import { type FastifyReply, type FastifyRequest } from 'fastify'
-import FindNeighborByIDUseCase from '../../../neighbor/aplication/usecases/find-by-id.usecase'
+import FindNeighborByIDUseCase from '../../../neighbor/application/usecases/find-by-id.usecase'
 import type NeighborRepository from '../../../neighbor/domain/repositories/neighbor.repository'
 import SchemaValidator from '../../../shared/infrastructure/middlewares/zod-schema-validator.middleware'
 import HandleHTTPResponse from '../../../shared/utils/http.reply.util'
 import { GetURLParams } from '../../../shared/utils/http.request.util'
-import FindWasteCategoryByIDUseCase from '../../../waste-category/aplication/usecases/find-by-id.usecase'
+import FindWasteCategoryByIDUseCase from '../../../waste-category/application/usecases/find-by-id.usecase'
 import type WasteCategoryRepository from '../../../waste-category/domain/repositories/waste-category.repository'
 import RegisterWasteTransactionDetailUseCase from '../../../waste-transaction-detail/application/usecases/register.usecase'
 import type WasteTransactionDetailRepository from '../../../waste-transaction-detail/domain/repositories/waste-transaction-detail.repository'
