+++ src/waste/infrastructure/handlers/waste.handler.ts
@@ -2,7 +2,7 @@ import { type FastifyReply, type FastifyRequest } from 'fastify'
 import SchemaValidator from '../../../shared/infrastructure/middlewares/zod-schema-validator.middleware'
 import HandleHTTPResponse from '../../../shared/utils/http.reply.util'
 import { GetURLParams } from '../../../shared/utils/http.request.util'
-import FindWasteCategoryByIDUseCase from '../../../waste-category/aplication/usecases/find-by-id.usecase'
+import FindWasteCategoryByIDUseCase from '../../../waste-category/application/usecases/find-by-id.usecase'
 import type WasteCategoryRepository from '../../../waste-category/domain/repositories/waste-category.repository'
 import FindWasteByIDUseCase from '../../application/usecases/find-by-id.usecase'
 import RegisterWasteUseCase from '../../application/usecases/register.usecase'
